+++ local_libs/lib-cloud-proxy/gradle.properties
@@ -1,5 +1,5 @@
-revision = 1.4.1-SNAPSHOT
-kotlinVersion=1.7.0
+revision = 1.0.0-SNAPSHOT
+kotlinVersion=1.7.20
 micronautVersion=3.5.3
 awsSDKVersion=2.17.234
