+++ local_libs/lib-cloud-proxy/src/main/kotlin/gov/cdc/dex/cloud/ProviderMeta.kt
@@ -1,0 +1,13 @@
+package gov.cdc.dex.cloud
+
+/**
+ * Intended to provide access to any provider details (which provider, region, account, etc.)
+ */
+interface ProviderMeta {
+    fun provider(): Providers
+}
+
+enum class Providers {
+    AWS,
+    Azure
+}
