+++ local_libs/lib-cloud-proxy/src/main/kotlin/gov/cdc/dex/cloud/messaging/azure/ASQMessage.kt
@@ -1,0 +1,18 @@
+package gov.cdc.dex.cloud.messaging.azure
+
+import com.jayway.jsonpath.JsonPath
+import gov.cdc.dex.cloud.messaging.CloudMessage
+
+class ASQMessage (
+    id: String,
+    recipientHandle: String,
+    body: String,
+    queue: String
+) : CloudMessage(id, recipientHandle, body, queue) {
+
+    private val ASQ_KEY_PATH:String = "data.url"
+
+    override fun key(): String {
+        return JsonPath.parse(this.body).read(ASQ_KEY_PATH, String::class.java)
+    }
+}
