--- local_libs/lib-cloud-proxy/src/main/kotlin/gov/cdc/ncezid/cloud/ProviderMeta.kt
@@ -1,13 +1,0 @@
-package gov.cdc.ncezid.cloud
-
-/**
- * Intended to provide access to any provider details (which provider, region, account, etc.)
- */
-interface ProviderMeta {
-    fun provider(): Providers
-}
-
-enum class Providers {
-    AWS,
-    Azure
-}
