--- local_libs/lib-cloud-proxy/src/main/kotlin/gov/cdc/ncezid/cloud/messaging/CloudMessage.kt
@@ -1,20 +1,0 @@
-package gov.cdc.ncezid.cloud.messaging
-
-abstract class CloudMessage(
-    val id: String,
-    val recipientHandle: String,
-    val body: String,
-    val queue: String
-) {
-    abstract fun key(): String
-
-    override fun toString(): String {
-        return """CloudMessage(
-            |  id='$id', 
-            |  recipientHandle='$recipientHandle', 
-            |  body='$body', 
-            |  key='${key()}',
-            |  queue='$queue'
-            |)""".trimMargin()
-    }
-}
